requirements.txt
Django==3.0.7
django-crispy-forms==1.9.1
pytz==2018.7
rpy2==3.3.5
pandas==1.0.5
django-multiselectfield==0.1.12
django-picklefield==3.0.1
celery-progress
celery==4.4.6
rabbitmq==0.2.0
amqp==2.6.0
